[SPARK-42944][PYTHON][FOLLOW-UP] Rename tests from foreachBatch to foreach_batch
This PR proposes to rename tests from foreachBatch to foreach_batch.

Non-API should follow snake_naming rule per PEP 8.

No, dev-only.

CI in this PR should test it out.

No.

Closes #42675 from HyukjinKwon/pyspark-connect.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
(cherry picked from commit 84a65bd1cf95775c4210c6dac8026551fd9d150f)
