+++ python/pyspark/sql/tests/connect/streaming/test_parity_foreach_batch.py
@@ -17,23 +17,23 @@
 
 import unittest
 
-from pyspark.sql.tests.streaming.test_streaming_foreachBatch import StreamingTestsForeachBatchMixin
+from pyspark.sql.tests.streaming.test_streaming_foreach_batch import StreamingTestsForeachBatchMixin
 from pyspark.testing.connectutils import ReusedConnectTestCase
 
 
 class StreamingForeachBatchParityTests(StreamingTestsForeachBatchMixin, ReusedConnectTestCase):
     @unittest.skip("SPARK-44463: Error handling needs improvement in connect foreachBatch")
-    def test_streaming_foreachBatch_propagates_python_errors(self):
-        super().test_streaming_foreachBatch_propagates_python_errors
+    def test_streaming_foreach_batch_propagates_python_errors(self):
+        super().test_streaming_foreach_batch_propagates_python_errors()
 
     @unittest.skip("This seems specific to py4j and pinned threads. The intention is unclear")
-    def test_streaming_foreachBatch_graceful_stop(self):
-        super().test_streaming_foreachBatch_graceful_stop()
+    def test_streaming_foreach_batch_graceful_stop(self):
+        super().test_streaming_foreach_batch_graceful_stop()
 
 
 if __name__ == "__main__":
     import unittest
-    from pyspark.sql.tests.connect.streaming.test_parity_foreachBatch import *  # noqa: F401,E501
+    from pyspark.sql.tests.connect.streaming.test_parity_foreach_batch import *  # noqa: F401,E501
 
     try:
         import xmlrunner  # type: ignore[import]
